other/requirements.txt
numpy
grpcio
h5py
protobuf